+++ src/LSST/M1M3/SS/ForceComponents/ElevationForceComponent.cpp
@@ -40,7 +40,7 @@ using namespace LSST::M1M3::SS;
 ElevationForceComponent::ElevationForceComponent(
         ForceActuatorApplicationSettings* forceActuatorApplicationSettings)
         : ForceComponent("Elevation", ForceActuatorSettings::instance().ElevationComponentSettings) {
-    _safetyController = Model::get().getSafetyController();
+    _safetyController = Model::instance().getSafetyController();
     _forceActuatorApplicationSettings = forceActuatorApplicationSettings;
     _forceSetpointWarning = M1M3SSPublisher::instance().getEventForceSetpointWarning();
     _appliedElevationForces = M1M3SSPublisher::instance().getAppliedElevationForces();